bpf: rename VTEP_MAP to cilium_vtep_map, move to vtep.h
