+++ bpf/node_config.h
@@ -140,7 +140,6 @@ volatile const __u64 __config_ROUTER_IP_2;
 #define RATELIMIT_MAP test_cilium_ratelimit
 #define RATELIMIT_METRICS_MAP test_cilium_ratelimit_metrics
 #define TUNNEL_MAP test_cilium_tunnel_map
-#define VTEP_MAP test_cilium_vtep_map
 #define LB6_REVERSE_NAT_MAP test_cilium_lb6_reverse_nat
 #define LB6_SERVICES_MAP_V2 test_cilium_lb6_services
 #define LB6_BACKEND_MAP test_cilium_lb6_backends
